+++ package.json
@@ -55,6 +55,7 @@
     "@pretextbook/completions": "^0.0.3",
     "@pretextbook/format": "^0.0.5",
     "@pretextbook/latex-pretext": "0.0.8",
+    "@pretextbook/remark-pretext": "^0.0.3",
     "@tailwindcss/vite": "^4.1.18",
     "@tiptap/extension-code-block": "^3.14.0",
     "@tiptap/extension-hard-break": "^3.14.0",
